Decomposed 'workspace section' component
